refactor: change nextauthoptions folder structure for project build
